Some code clean up
